Handle files over 2GB (#1065)
* Handle files over 2GB

* Use Guava's checkArgument

+++ gcloud-java-core/src/main/java/com/google/cloud/ReadChannel.java
@@ -38,7 +38,7 @@ public interface ReadChannel extends ReadableByteChannel, Closeable, Restorable<
   @Override
   void close();
 
-  void seek(int position) throws IOException;
+  void seek(long position) throws IOException;
 
   /**
    * Sets the minimum size that will be read by a single RPC.
